Schedule & Clock Separation
- Moved dependency of schedule and clock away from each other, as much as possible
- May be further areas of improvements, will revisit tomorrow during school day
